requirements.txt
tweepy
pytz
django-scheduler
PIL
